новый функционал

+++ src/main/kotlin/ru/mai/coursework/entity/Payment.kt
@@ -7,7 +7,7 @@ import java.time.LocalDateTime
 data class Payment(
     val paymentId: Int,
     val bookingId: Int,
-    val amount: BigDecimal,
+    val amount: Float,
     val paymentTime: LocalDateTime,
     val paymentMethod: String?,
     val status: PaymentStatus
@@ -17,7 +17,7 @@ fun ResultSet.toPayment(): Payment {
     return Payment(
         paymentId = getInt("payment_id"),
         bookingId = getInt("booking_id"),
-        amount = getBigDecimal("amount"),
+        amount = getFloat("amount"),
         paymentTime = getTimestamp("payment_time").toLocalDateTime(),
         paymentMethod = getString("payment_method"),
         status = PaymentStatus.valueOf(getString("status").uppercase())
